Continue File & Facility refactor
