+++ FMS.Domain/Repositories/IFacilityRepository.cs
@@ -21,6 +21,7 @@ namespace FMS.Domain.Repositories
         Task UndeleteFacilityAsync(Guid id);
         Task<bool> FacilityNumberExists(string facilityNumber, Guid? ignoreId = null);
         Task<bool> FileLabelExists(string fileLabel);
+        Task<int> GetNextSequenceForCountyAsync(int countyId);
 
         // Retention Records
         Task<bool> RetentionRecordExistsAsync(Guid id);
